third_party/WebKit/LayoutTests/fast/css/fontface-properties.html

<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
<style>
@font-face {
  font-family: 'Ahem';
  src: url(../../resources/Ahem.ttf);
  font-style: italic;
  font-weight: 300;
  unicode-range: U+0-3FF;
  font-variant: small-caps;
  -webkit-font-feature-settings: "dlig" 1;
  /* font-stretch property is not supported */
}
</style>
<script>
description('Test getting/setting FontFace properties.');

function runTests() {
    ahemFace = document.fonts.match('italic small-caps 10px Ahem')[0];
    shouldBeEqualToString('ahemFace.family', 'Ahem');
    shouldBeEqualToString('ahemFace.style', 'italic');
    shouldBeEqualToString('ahemFace.weight', '300');
    shouldBeEqualToString('ahemFace.unicodeRange', 'U+0-3FF');
    shouldBeEqualToString('ahemFace.variant', 'small-caps');
    shouldBeEqualToString('ahemFace.featureSettings', "'dlig' 1");

    debug('');
    defaultFace = new FontFace('defaultFace', 'local(foo)', {});
    shouldBeEqualToString('defaultFace.family', 'defaultFace');
    shouldBeEqualToString('defaultFace.style', 'normal');
    shouldBeEqualToString('defaultFace.weight', 'normal');
    shouldBeEqualToString('defaultFace.stretch', 'normal');
    shouldBeEqualToString('defaultFace.unicodeRange', 'U+0-10FFFF');
    shouldBeEqualToString('defaultFace.variant', 'normal');
    shouldBeEqualToString('defaultFace.featureSettings', 'normal');

    debug('');
    constructedFace = new FontFace('constructedFace', 'local(bar)', {
        'style': 'oblique',
        'weight': 'bold',
        'unicodeRange': 'U+100-1FF, U+ABCD',
        'variant': 'small-caps',
        'featureSettings': "'liga' 0"
    });
    shouldBeEqualToString('constructedFace.family', 'constructedFace');
    shouldBeEqualToString('constructedFace.style', 'oblique');
    shouldBeEqualToString('constructedFace.weight', 'bold');
    shouldBeEqualToString('constructedFace.unicodeRange', 'U+100-1FF, U+ABCD');
    shouldBeEqualToString('constructedFace.variant', 'small-caps');
    shouldBeEqualToString('constructedFace.featureSettings', "'liga' 0");

    debug('');
    modifiedFace = new FontFace('unmodified', 'local(baz)', {});
    modifiedFace.family = 'modified';
    modifiedFace.style = 'italic';
    modifiedFace.weight = 900;
    modifiedFace.unicodeRange = 'U+0-3FF';
    modifiedFace.variant = 'small-caps';
    modifiedFace.featureSettings = "'dlig' 1, 'liga' 0";
    shouldBeEqualToString('modifiedFace.family', 'modified');
    shouldBeEqualToString('modifiedFace.style', 'italic');
    shouldBeEqualToString('modifiedFace.weight', '900');
    shouldBeEqualToString('modifiedFace.unicodeRange', 'U+0-3FF');
    shouldBeEqualToString('modifiedFace.variant', 'small-caps');
    shouldBeEqualToString('modifiedFace.featureSettings', "'dlig' 1, 'liga' 0");

    debug('');
    var syntaxError = "'SyntaxError: An invalid or illegal string was specified.'";
    shouldThrow("new FontFace('test', 'invalid_src', {})", syntaxError);
    face = new FontFace('test', 'local(foo)', {});
    shouldThrow("face.style = ''", syntaxError);
    shouldThrow("face.weight = 'a'", syntaxError);
    shouldThrow("face.unicodeRange = 'U+'", syntaxError);
    shouldThrow("face.variant = '???'", syntaxError);
    shouldThrow("face.featureSettings = null", syntaxError);
}

if (document.fonts)
    runTests();
else {
    testFailed('document.fonts does not exist');
}
</script>
</head>
<body>
  <script src="../js/resources/js-test-post.js"></script>
</body>
</html>
